requirements.txt
langgraph 
langchain 
langchain-groq 
flask 
pydantic 
python-dotenv 
arxiv 
wikipedia 
langchain_community